v0.2-b1 new config parser
This introduces a new configuration parser that allows more than
just the PATH element. We implement a LOG element with a file: prefix
as appending a log message to a file.  This is not very scalable but
it is the first thing to implement.  More busy wrappers will require
other log mechanisms which are async and non-blocking.